# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School ID App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/register-school:
    post:
      operationId: registerSchool
      tags: [auth]
      summary: Register a new school and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSchoolBody"
      responses:
        "201":
          description: School and admin created, token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "409":
          description: Slug, code, or username already taken
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
...
  /behavior/logs:
    get:
      operationId: listBehaviorLogs
      tags: [behavior]
      summary: List behavior logs
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of behavior logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BehaviorLog"
    post:
      operationId: createBehaviorLog
      tags: [behavior]
      summary: Log a behavior event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBehaviorLogBody"
      responses:
        "201":
          description: Behavior logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BehaviorLog"
  /behavior/logs/{id}:
    patch:
      operationId: updateBehaviorLog
      tags: [behavior]
      summary: Update a behavior log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBehaviorLogBody"
      responses:
        "200":
          description: Behavior log updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BehaviorLog"
    delete:
      operationId: deleteBehaviorLog
      tags: [behavior]
      summary: Delete a behavior log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /behavior/categories:
    get:
      operationId: listBehaviorCategories
      tags: [behavior]
      summary: List behavior categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BehaviorCategory"
    post:
      operationId: createBehaviorCategory
      tags: [behavior]
      summary: Create a behavior category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBehaviorCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BehaviorCategory"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List staff/admin users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a staff/admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        isActive:
          type: string
        schoolId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - username
        - firstName
        - lastName
        - role
        - isActive
        - schoolId
        - createdAt
    RegisterSchoolBody:
      type: object
      properties:
        schoolName:
          type: string
        schoolSlug:
          type: string
        schoolCode:
          type: string
        adminUsername:
          type: string
        adminPassword:
          type: string
        adminFirstName:
          type: string
        adminLastName:
          type: string
        contactEmail:
          type: string
      required:
        - schoolName
        - schoolSlug
        - schoolCode
        - adminUsername
        - adminPassword
        - adminFirstName
        - adminLastName
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
      required:
        - username
        - password
        - firstName
        - lastName
        - role
...
    UpdateBehaviorLogBody:
      type: object
      properties:
        type:
          type: string
          enum: [merit, demerit]
        categoryId:
          type: ["integer", "null"]
        points:
          type: integer
        note:
          type: ["string", "null"]
    BehaviorLog:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
        points:
          type: integer
        note:
          type: ["string", "null"]
        loggedById:
          type: ["integer", "null"]
        createdAt:
          type: string
        categoryName:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - type
        - points
        - createdAt
    CreateBehaviorLogBody:
      type: object
      properties:
        studentId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        type:
          type: string
        points:
          type: integer
        note:
          type: ["string", "null"]
      required:
        - studentId
        - type
        - points
    BehaviorCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        points:
          type: integer
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - points
        - createdAt
    CreateBehaviorCategoryBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        points:
          type: integer
        description:
          type: ["string", "null"]
      required:
        - name
        - type
        - points
...