# lib/api-zod/src/generated/types/registerSchoolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School ID App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterSchoolBody {
  schoolName: string;
  schoolSlug: string;
  schoolCode: string;
  adminUsername: string;
  adminPassword: string;
  adminFirstName: string;
  adminLastName: string;
  contactEmail?: string;
}