# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School ID App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  username: string;
  firstName: string;
  lastName: string;
  role: string;
  isActive: string;
  schoolId: number;
  createdAt: string;
}